Ardu copter 3.2.1

+++ libraries/GCS_MAVLink/include/mavlink/v1.0/checksum.h
@@ -36,6 +36,7 @@ static inline void crc_accumulate(uint8_t data, uint16_t *crcAccum)
 }
 #endif
 
+
 /**
  * @brief Initiliaze the buffer for the X.25 CRC
  *
@@ -64,6 +65,7 @@ static inline uint16_t crc_calculate(const uint8_t* pBuffer, uint16_t length)
         return crcTmp;
 }
 
+
 /**
  * @brief Accumulate the X.25 CRC by adding an array of bytes
  *
@@ -73,16 +75,13 @@ static inline uint16_t crc_calculate(const uint8_t* pBuffer, uint16_t length)
  * @param data new bytes to hash
  * @param crcAccum the already accumulated checksum
  **/
-static inline void crc_accumulate_buffer(uint16_t *crcAccum, const char *pBuffer, uint8_t length)
+static inline void crc_accumulate_buffer(uint16_t *crcAccum, const char *pBuffer, uint16_t length)
 {
 	const uint8_t *p = (const uint8_t *)pBuffer;
 	while (length--) {
                 crc_accumulate(*p++, crcAccum);
         }
 }
-
-
-
 
 #endif /* _CHECKSUM_H_ */
